pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "brownfield-cartographer"
version = "0.1.0"
description = "Codebase intelligence system for brownfield data platforms (The Brownfield Cartographer)."
authors = [
  { name = "10 Academy Fellow" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.6.0",
  "networkx>=3.2.0",
  "sqlglot>=25.0.0",
  "tree-sitter>=0.23.0",
  "tree-sitter-language-pack>=0.13.0",
  "typer>=0.12.0",
  "gitpython>=3.1.40",
  "pyyaml>=6.0",
]

[project.scripts]
cartographer = "src.cli:app"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.uv]
dev-dependencies = []

